Created exsta wide shared folder and moved domain entities into it, also renamed previous shared to backend-shared for clarity
