+++ Exsta/Exsta.sln
@@ -24,7 +24,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".Shared", ".Shared", "{4BA60918-A374-4671-A1EB-D4A9F6FE546D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shared", "Shared\Shared.csproj", "{7C9FFEA1-B770-4CD5-A58E-044268ED9EB0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Backend-Shared", "Shared\Backend-Shared.csproj", "{7C9FFEA1-B770-4CD5-A58E-044268ED9EB0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Exsta-Shared", "..\Exsta-Shared\Exsta-Shared.csproj", "{07A60AA8-386F-41C5-87B0-137B1A314657}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -44,6 +46,10 @@ Global
 		{7C9FFEA1-B770-4CD5-A58E-044268ED9EB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7C9FFEA1-B770-4CD5-A58E-044268ED9EB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7C9FFEA1-B770-4CD5-A58E-044268ED9EB0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{07A60AA8-386F-41C5-87B0-137B1A314657}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{07A60AA8-386F-41C5-87B0-137B1A314657}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{07A60AA8-386F-41C5-87B0-137B1A314657}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{07A60AA8-386F-41C5-87B0-137B1A314657}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -52,6 +58,7 @@ Global
 		{107AE5CF-BE4F-470B-9E18-6EA6EEC96FC3} = {38B13A26-7F27-4BEA-939A-A97FBA21AE99}
 		{5AEC2B44-A254-45E1-A503-CE4C3B27D278} = {4F16DD18-8964-4088-B085-EFC21DA956C2}
 		{7C9FFEA1-B770-4CD5-A58E-044268ED9EB0} = {4BA60918-A374-4671-A1EB-D4A9F6FE546D}
+		{07A60AA8-386F-41C5-87B0-137B1A314657} = {4BA60918-A374-4671-A1EB-D4A9F6FE546D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7F00C589-974F-4187-BF0C-CC10607B6A73}
